Put running_process{,_mut} in crate::system
